requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coverage==5.5
flake8==4.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
grpcio==1.40.0
grpcio-tools==1.40.0
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pluggy==1.0.0
protobuf==3.18.0
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytest-mock==3.6.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
